+++ integration_tests/codegen_fail/fail/scalar/mutliple_named_fields.stderr
@@ -1,4 +1,4 @@
-error: GraphQL scalar expected exactly 1 field, e.g., Test { test: i32 }or all `resolve`, `from_input_value` and `from_str` functions
+error: GraphQL scalar expected exactly 1 field, e.g., Test { test: i32 } or all custom resolvers
  --> fail/scalar/mutliple_named_fields.rs:4:1
   |
 4 | / struct Scalar {
