+++ integration_tests/codegen_fail/fail/scalar/mutliple_unnamed_fields.stderr
@@ -1,4 +1,4 @@
-error: GraphQL scalar expected exactly 1 field, e.g., Test(i32)or all `resolve`, `from_input_value` and `from_str` functions
+error: GraphQL scalar expected exactly 1 field, e.g., Test(i32) or all custom resolvers
  --> fail/scalar/mutliple_unnamed_fields.rs:4:1
   |
 4 | struct Scalar(i32, i32);
